Bug-Fix: Profile : Timeline Activity - Cough Test items missing time of day time stamp.  Users can do Cough tests multiple times in one day so there should be time indication
